Refactor via agents.md.
