+++ lang/en.ts
@@ -6,12 +6,13 @@ const lang = {
   settings: "Simple Focus Settings",
   showRightClickMenu: "Show right-click menu option",
   showFileExplorerIcon: "Show file explorer icon",
-  focusLevel: "Focus level",
+  focusLevel: "Command focus level",
   focusLevelFile: "Current file only",
   focusLevelParent: "Parent folder",
   focusLevelGrandparent: "Grandparent folder",
   focusLevelGreatGrandparent: "Great grandparent folder",
   focusLevelCustom: "Specific folder",
+  focusLevelDesc: "Determines what to focus when using the toggle command or file explorer icon. Right-click menu always focuses the clicked file/folder.",
   customFolderPath: "Custom folder path",
   customFolderPathDesc: "Enter a folder path (folder/subfolder). This folder will be focused regardless of which file is open.",
 }
