+++ lang/zh.ts
@@ -6,12 +6,13 @@ const lang = {
   settings: "简单专注设置",
   showRightClickMenu: "显示右键菜单选项",
   showFileExplorerIcon: "显示文件浏览器图标",
-  focusLevel: "专注级别",
+  focusLevel: "命令专注级别",
   focusLevelFile: "仅当前文件",
   focusLevelParent: "父文件夹",
   focusLevelGrandparent: "祖父文件夹",
   focusLevelGreatGrandparent: "曾祖父文件夹",
   focusLevelCustom: "特定文件夹",
+  focusLevelDesc: "确定使用切换命令或文件浏览器图标时要聚焦的内容。右键菜单始终聚焦单击的文件/文件夹。",
   customFolderPath: "自定义文件夹路径",
   customFolderPathDesc: "输入文件夹路径（文件夹/子文件夹）。无论打开哪个文件，都会聚焦此文件夹。",
 }
