+++ manifest.json
@@ -1,7 +1,7 @@
 {
 	"id": "simple-focus",
 	"name": "Simple Focus",
-	"version": "0.0.5",
+	"version": "0.0.6",
 	"minAppVersion": "0.15.0",
 	"description": "Allows you to focus on a specific file or folder",
 	"author": "linqing",
