+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "obsidian-simple-focus-plugin",
-	"version": "0.0.5",
+	"version": "0.0.6",
 	"description": "This plugin allows you to focus on a specific file or folder",
 	"main": "main.js",
 	"scripts": {
