+++ versions.json
@@ -3,5 +3,6 @@
 	"0.0.2": "0.15.0",
 	"0.0.3": "0.15.0",
 	"0.0.4": "0.15.0",
-	"0.0.5": "0.15.0"
+	"0.0.5": "0.15.0",
+	"0.0.6": "0.15.0"
 }
